requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
channels==4.3.1
channels-redis==4.3.0
django-cors-headers==4.9.0
redis==7.0.1
PyJWT==2.10.1
msgpack==1.1.2
